Move OPS CRD into a separate package to cut dependency
