+++ pkg/ops/crd/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["crd.go"],
+    importpath = "github.com/atlassian/voyager/pkg/ops/crd",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/apis/ops:go_default_library",
+        "//pkg/apis/ops/v1:go_default_library",
+        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+    ],
+)
